refactor(adapters): move legacy outbound-log/worklog inside discord adapter (#1787)
Relocates the legacy JSONL outbound-log + worklog concern out of
cortex.ts and behind a single adapter seam, so core no longer reaches
into discord.js internals — unblocking the discord adapter extraction
(S12, epic #1784).

- setupOutboundLog + helpers → src/adapters/discord/outbound-log.ts,
  exposed as attachLegacyOutboundLog() (re-exported via the discord
  index); cortex.ts + surface-adapter-boot.ts call the seam, not
  getClient()/formatEventForDiscord.
- worklog-manager/worklog-formatter/event-utils moved into
  src/adapters/discord/ (event-utils's only two importers moved with
  it; distinct from the untouched src/common/event-utils.ts).
- Behaviour-preserving: JSONL polling + router worklog-surface
  registration unchanged (no MIG-7.2d renderer redesign).

Gates: tsc 0 · bun test src/adapters/discord src/runner src/bus
2407 pass/0 fail · eslint 0. Rebased on origin/main.
