+++ src/cortex.ts
@@ -18,11 +18,10 @@
 import "./bootstrap/ws-transport";
 
 import { Command } from "commander";
-import { existsSync, writeFileSync, readFileSync, unlinkSync, mkdirSync, readdirSync } from "fs";
+import { existsSync, writeFileSync, readFileSync, unlinkSync, mkdirSync } from "fs";
 import { join, dirname, isAbsolute } from "path";
 import { homedir } from "os";
 import { parse as parseYaml } from "yaml";
-import type { TextChannel } from "discord.js";
 
 import {
   loadConfigWithAgents,
@@ -131,7 +130,7 @@ import {
   type Envelope,
 } from "./bus/myelin/envelope-validator";
 
-import { DiscordAdapter } from "./adapters/discord";
+import { attachLegacyOutboundLog } from "./adapters/discord";
 import { createRenderer, type Renderer } from "./renderers";
 import { RendererSchema, deriveStackId, DEFAULT_STREAM_MAX_BYTES } from "./common/types/cortex-config";
 import type {
@@ -303,7 +302,6 @@ import type {
 } from "./surface/mc/local-aggregation/sibling-db-reader";
 import type { SiblingStackDescriptor } from "./surface/mc/local-aggregation/sibling-discovery";
 import { createProbeResponder, type ProbeResponder } from "./bus/probe-responder";
-import { WorklogManager } from "./runner/worklog-manager";
 
 // #752 — the `network` + `provision-stack` subcommand dispatchers. Registered
 // as passthrough commander subcommands below so `cortex network join …` and
@@ -341,9 +339,6 @@ import type {
 } from "./common/registry";
 import { resolveBootFederatedPeers } from "./common/registry/resolve-federated-peers-boot";
 import type { NetworkRosterProvider } from "./common/registry/resolve-federated-peers";
-import { JsonlReader } from "./taps/cc-events/lib/jsonl-reader";
-import { PublishedEventSchema } from "./taps/cc-events/hooks/lib/event-types";
-import { formatEventForDiscord } from "./adapters/discord/event-formatter";
 import {
   startGithubWebhookReceiver,
   type GithubWebhookReceiverHandle,
@@ -3199,7 +3194,7 @@ export async function startCortex(
     trustResolver,
     inboundWithGateBridge,
     disableOutboundPoller: options.disableOutboundPoller ?? false,
-    setupOutboundLog,
+    setupOutboundLog: attachLegacyOutboundLog,
     adapters,
     adapterCleanup,
     liveSurfaces,
@@ -5999,104 +5994,6 @@ async function resolveReviewProvisioningJsm(
     );
     return null;
   }
-}
-
-/** Subscribe a Discord adapter to the published-events JSONL stream so legacy
- *  `#agent-log` and per-task worklog threads keep working. Mirrors grove-bot's
- *  `setupOutboundLog`. The dispatch.task.* projection has migrated to
- *  `worklog-manager.surfaceConfig` (router-driven); this function runs the
- *  legacy direct-call path while MIG-7.2d Renderer cutover is pending. */
-function setupOutboundLog(
-  discordAdapter: DiscordAdapter,
-  instance: import("./common/types/config").DiscordInstance,
-  config: AgentConfig,
-  router: SurfaceRouter,
-  systemEventSource: SystemEventSource,
-): (() => void) | null {
-  const eventsDir = config.paths.publishedEventsDir.replace(/^~/, process.env.HOME ?? "~");
-  const client = discordAdapter.getClient();
-  if (!client) {
-    console.log("cortex: discord client not available for outbound log");
-    return null;
-  }
-
-  let pollInterval: ReturnType<typeof setInterval> | null = null;
-
-  // discord.js v15 prep — see comment in adapters/discord/client.ts.
-  client.on("clientReady", () => {
-    if (!existsSync(eventsDir)) {
-      console.log(`cortex: published events dir not found (${eventsDir}), skipping outbound`);
-      return;
-    }
-
-    const reader = new JsonlReader();
-    reader.skipAllToEnd(eventsDir);
-    const logChannelId = instance.logChannelId;
-    const guildId = instance.guildId;
-    const postedEventIds = new Set<string>();
-
-    let worklog: WorklogManager | null = null;
-    if (instance.worklogChannelId) {
-      worklog = new WorklogManager(client, instance.worklogChannelId);
-      console.log(`cortex: worklog enabled → channel ${instance.worklogChannelId}`);
-      // Register the worklog manager's `dispatch.task.*` surface so the
-      // bus-driven path projects into the same threads as the JSONL path.
-      router.register(
-        worklog.surfaceConfig({
-          principal: systemEventSource.principal,
-          adapterId: `worklog-${discordAdapter.instanceId}`,
-        }),
-      );
-    }
-
-    const processFile = async (path: string) => {
-      const events = reader.readNew(path);
-      if (events.length > 0) {
-        console.log(`cortex: processing ${events.length} event(s) from ${path.split("/").pop()}`);
-      }
-      for (const raw of events) {
-        try {
-          const event = PublishedEventSchema.parse(raw);
-          if (postedEventIds.has(event.event_id)) continue;
-          postedEventIds.add(event.event_id);
-          if (worklog) await worklog.handleEvent(event);
-          if (instance.enableAgentLog) {
-            const formatted = formatEventForDiscord(event);
-            if (!formatted) continue;
-            const guild = client.guilds.cache.get(guildId);
-            const channel =
-              (guild?.channels.cache.get(logChannelId) as TextChannel | null)
-              ?? ((await client.channels.fetch(logChannelId).catch(() => null)) as TextChannel | null);
-            if (channel && "send" in channel) {
-              await channel.send(formatted);
-            } else {
-              console.error(`cortex: could not resolve log channel ${logChannelId} — check bot permissions and channel ID`);
-            }
-          }
-        } catch (err) {
-          console.error("cortex: outbound error:", err instanceof Error ? err.message : err);
-        }
-      }
-    };
-
-    pollInterval = setInterval(() => {
-      try {
-        const files = readdirSync(eventsDir).filter((f: string) => f.endsWith(".jsonl"));
-        for (const file of files) void processFile(join(eventsDir, file));
-      } catch (err) {
-        console.error("cortex: poll error:", err instanceof Error ? err.message : String(err));
-      }
-    }, 2000);
-
-    console.log(`cortex: polling ${eventsDir} for outbound events (every 2s)`);
-  });
-
-  return () => {
-    if (pollInterval !== null) {
-      clearInterval(pollInterval);
-      pollInterval = null;
-    }
-  };
 }
 
 // CLI
